Minor tweaks
Decided to begin with a 3-party checker.
